PdfWriter: Added support for updating files with broken XRef sections
Spam a warning if attempting to do so. This effectively allows to sign some broken PDF documents.
This also fixes signing the same document more times.

+++ src/podofo/main/PdfMemDocument.h
@@ -195,7 +195,7 @@ private:
     PdfVersion m_InitialVersion;
     bool m_HasXRefStream;
     size_t m_MagicOffset;
-    int64_t m_PrevXRefOffset;
+    size_t m_PrevXRefOffset;
     std::unique_ptr<PdfEncryptSession> m_Encrypt;
     std::shared_ptr<InputStreamDevice> m_device;
 };
